Extract shared event models
